Added isrequired to the field type filter and runtime editors and worked on the management and editor pages of GenericBusinessEntity
git-tfs-id: [http://developmentvlan:8585/tfs/vanrise.collection]$/;C31266

+++ Vanrise/Code/Vanrise.GenericData.Entities/GenericBusinessEntity/GenericBusinessEntityInfo.cs
@@ -8,7 +8,7 @@ namespace Vanrise.GenericData.Entities
 {
     public class GenericBusinessEntityInfo
     {
-        public int GenericBusinessEntityId { get; set; }
+        public long GenericBusinessEntityId { get; set; }
         public string Name { get; set; }
     }
 }
